Create PrimeNumbersAndMersenneNumbers
The first function determines if a number is prime by returning True or False.
The second function accepts 2 attributes that serve as the range for a list of mersenne numbers to be printed from. It will first iterate over the range of numbers to determine if they are prime before printing out a list of mersenne numbers.